src/Sensor/SDS011.API/requirements.txt
Flask~=1.1
pyserial